Update error handling (#12)
* Update error page

* Abstract api call to handle errors globally

* Fix bug in delete methods

* Update all services with abstracted function

* Move services to separate folder

* Remove local handlers for 500

+++ app/Http/Controllers/NoticeController.php
@@ -79,8 +79,10 @@ class NoticeController extends Controller
         $image = $notice->image;
 
         $notice->delete();
-        Storage::delete($image->path);
-        $image->delete();
+        if ($image) {
+            Storage::delete($image->path);
+            $image->delete();
+        }
     }
 
     public function UpdateUuids()
