+++ app/Http/Controllers/ScreeningController.php
@@ -142,8 +142,10 @@ class ScreeningController extends Controller
         $image = $screening->image;
 
         $screening->delete();
-        Storage::delete($image->path);
-        $image->delete();
+        if ($image) {
+            Storage::delete($image->path);
+            $image->delete();
+        }
     }
 
     public function UpdateUuids()
